Move roster sorting logic to backend
