+++ app/features/team/routes/t.$customUrl.tsx
@@ -87,17 +87,6 @@ export const handle: SendouRouteHandle = {
 export default function TeamPage() {
 	const { team } = useLoaderData<typeof loader>();
 
-	// Sort member by their role, in the order of TEAM_MEMBER_ROLES
-	team.members.sort((a, b) => {
-		const aIndex = TEAM_MEMBER_ROLES.indexOf(
-			a.role as (typeof TEAM_MEMBER_ROLES)[number],
-		);
-		const bIndex = TEAM_MEMBER_ROLES.indexOf(
-			b.role as (typeof TEAM_MEMBER_ROLES)[number],
-		);
-		return aIndex - bIndex;
-	});
-
 	return (
 		<Main className="stack lg">
 			<div className="stack sm">
